requirements.txt
Flask
Flask-CORS
groq
requests
google-generativeai